+++ src/main/scala/com/github/gchudnov/mtg/internal/BasicRel.scala
@@ -189,75 +189,3 @@ transparent trait BasicRel[+T]:
    */
   final def equalsTo[T1 >: T](b: Interval[T1])(using bOrd: Ordering[Boundary[T1]]): Boolean =
     a.nonEmpty && b.nonEmpty && bOrd.equiv(a.left, b.left) && bOrd.equiv(a.right, b.right)
-
-// /**
-//  * Checks whether A is less-or-equal to B
-//  *
-//  * A ≤ B
-//  *
-//  * {{{
-//  *   a- <= b-
-//  *   a+ <= b+
-//  * }}}
-//  *
-//  * {{{
-//  *   - before         | b
-//  *   - meets          | m
-//  *   - overlaps       | o
-//  *   - starts         | s
-//  *   - is-finished-by | F
-//  *   - equal          | e
-//  * }}}
-//  */
-// final def isLessEqual[T1 >: T](b: Interval[T1])(using bOrd: Ordering[Boundary[T1]]): Boolean =
-//   a.nonEmpty && b.nonEmpty && bOrd.lteq(a.left, b.left) && bOrd.lteq(a.right, b.right)
-
-// /**
-//  * Checks whether A greater-than-or-equal-to B (Order Relation)
-//  *
-//  * A ≥ B
-//  *
-//  * {{{
-//  *   a- >= b-
-//  *   a+ >= b+
-//  *
-//  *   A ≥ B <=> r3 ∧ (¬r2 ∨ ¬r4)
-//  * }}}
-//  *
-//  * {{{
-//  *   - after            | B
-//  *   - is-met-by        | M
-//  *   - is-overlapped-by | O
-//  *   - finishes         | f
-//  *   - is-started-by    | S
-//  *   - equal            | e
-//  * }}}
-//  */
-// final def isGreaterEqual[T1 >: T](b: Interval[T1])(using bOrd: Ordering[Boundary[T1]]): Boolean =
-//   a.nonEmpty && b.nonEmpty && bOrd.gteq(a.left, b.left) && bOrd.gteq(a.right, b.right)
-
-// /**
-//  * Checks whether A succeeds B (Order Relation)
-//  *
-//  * A > B
-//  *
-//  * {{{
-//  *   - after | B
-//  * }}}
-//  */
-// final def isGreater[T1 >: T](b: Interval[T1])(using bOrd: Ordering[Boundary[T1]]): Boolean =
-//   a.nonEmpty && b.nonEmpty && bOrd.gt(a.left, b.left) && bOrd.gt(a.right, b.right)
-
-// /**
-//  * Precedes or IsEqual
-//  *
-//  * TODO: ADD TESTS
-//  *
-//  * A ≼ B
-//  *
-//  * {{{
-//  *   a+ <= b-
-//  * }}}
-//  */
-// final def isBeforeEqual[T1 >: T](b: Interval[T1])(using bOrd: Ordering[Boundary[T1]]): Boolean =
-//   a.nonEmpty && b.nonEmpty && bOrd.lt(a.right, b.left)
